Tests/LibWeb/Text/input/css/gradient-display-p3-linear-parsing.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    #foo {
        background-image: linear-gradient(in display-p3-linear, red, blue);
    }
</style>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
    });
</script>
