Tests/LibWeb/Text/expected/css/gradient-display-p3-linear-parsing.txt

#foo { background-image: linear-gradient(in display-p3-linear, red, blue); }
